text averaged embedding outside of api

+++ vectorizer_api/vectorizer/embed.py
@@ -34,6 +34,6 @@ def inference_glove_embedding(preprocessed_text, averaged_embedding=True):
 
 	# using float to ensure accuracy
 	# if averaged_embedding:
-	embedding_matrix = np.mean(embedding_matrix, axis=0, dtype=np.float64)
+	embedding_matrix = np.mean(embedding_matrix, axis=0)
 
 	return embedding_matrix
